Enum DocumentType moved to Database

+++ triss-core/src/main/kotlin/com/pp/trisscore/repository/IdentityDocumentRepository.kt
@@ -8,5 +8,5 @@ import reactor.core.publisher.Mono
 @Repository
 interface IdentityDocumentRepository : ReactiveCrudRepository<IdentityDocument,Long> {
 
-    fun findByEmployeeIdAndTypeAndNumber(employeeId:Long, type:Int, number:String) : Mono<IdentityDocument>
+    fun findByEmployeeIdAndTypeAndNumber(employeeId:Long, type:Long, number:String) : Mono<IdentityDocument>
 }
